fix gpu_acceleration (#40); remove torch import
