+++ pyproject.toml
@@ -20,6 +20,10 @@ typing_extensions = "^4.14.1"
 musicxml = "1.4"
 rapidocr-onnxruntime = "^1.4.4"
 onnxruntime="^1.22.1"
+requests="^2.32.5"
+
+[tool.poetry.group.gpu.dependencies]
+onnxruntime-gpu = "^1.22.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.5"
