+++ training/transformer/training_vocabulary.py
@@ -7,6 +7,7 @@ from homr.transformer.configs import default_config
 from homr.transformer.vocabulary import (
     EncodedSymbol,
     Vocabulary,
+    sort_token_chords
 )
 
 vocab = Vocabulary()
@@ -72,24 +73,6 @@ def _chord_to_str(chord: list[EncodedSymbol]) -> str:
                 list(lower_slurs_ties)
             )
     return str.join("&", [str(c) for c in annotation_resorted])
-
-
-def sort_token_chords(
-    symbols: list[EncodedSymbol], keep_chord_symbol: bool = False
-) -> list[list[EncodedSymbol]]:
-    chords: list[list[EncodedSymbol]] = []
-    is_in_chord = False
-    for symbol in symbols:
-        if symbol.rhythm == "chord":
-            is_in_chord = True
-        elif is_in_chord and len(chords) > 0:
-            if keep_chord_symbol:
-                chords[-1].append(EncodedSymbol("chord"))
-            chords[-1].append(symbol)
-            is_in_chord = False
-        else:
-            chords.append([symbol])
-    return chords
 
 
 def calc_ratio_of_tuplets(symbols: list[EncodedSymbol]) -> float:
